move demo to tests
* import xrange from six for python 3 compatibility
* fix pyrheliometric formula is from 1980
